app_template/requirements.txt
daovoice-sdk >=0.0.1
django >=1.11

